Delete the upstream/docs folder during provider builds
Upstream maintains 2 docs folders (`docs` and `website`). `website` is technically legacy,
but the TF registry draws documentation from it and we should too.

This should be fixed in the bridge (so we pull what the registry does), but this provides
an immediate fix.

I'll open an associated issue to track a bridge fix and the removal of this patch.