Remove docs patch (#595)
The docs patch existed because of
https://github.com/pulumi/pulumi-okta/issues/497.
https://github.com/pulumi/pulumi-terraform-bridge/issues/1609 was the
root cause, and it is now fixed. We can remove the docs patch.

We cannot transition back to a normal go dependency, since upstream
hasn't marked the major version in their go.mod. Ironically, since we
have the dependency downloaded we don't actually have to change
anything. I have added a empty patch to resolve this catch 22 scenario.